Conflicts:
	README.md

+++ src/dcamj/DcamAcquisitionListener.java
@@ -6,7 +6,7 @@ public interface DcamAcquisitionListener
 	void frameArrived(DcamAcquisition pDcamAquisition,
 										long pAbsoluteFrameIndex,
 										long pArrivalTimeStampInNanoseconds,
-										int pFrameIndexInBufferList,
+										long pFrameIndexInBufferList,
 										DcamFrame pDcamFrame);
 
 }
